feat: implement real qlib training with WebSocket progress tracking
- Implement actual qlib IC calculation in ModelTrainer
- Add WebSocket progress sync for training UI
- Add job cancellation support (DELETE /jobs/{job_id})
- Add model deletion for running/queued training
- Support float progress values (1 decimal precision)
- Add monthly date selector for training end date
- Fix page refresh losing training progress
- Fix timezone issue in training repository

+++ src/repositories/job.py
@@ -50,7 +50,7 @@ class JobRepository:
         self,
         job_id: str,
         status: str,
-        progress: int | None = None,
+        progress: float | None = None,
         message: str | None = None,
     ) -> Job | None:
         """更新任務狀態"""
